pdf_field_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pymupdf==1.24.14
opencv-python-headless==4.10.0.84
numpy==2.1.3
pytesseract==0.3.13
pillow==11.0.0
